requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
discord==2.3.2
discord-pretty-help==2.0.7
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
yarl==1.9.4
flask
discord-android==1.1.0